scripts/python/requirements.txt
click==7.0
faker==1.0.2
pandas==0.23.4
pip==19.0.2
python-decouple==3.1
pyodbc==4.0.25
SQLAlchemy==1.2.18
